requirements.txt
streamlit
pandas
requests
pandas-ta
plotly
python-dotenv
alpha-vantage